move quotes logic in About Component into own comoponent --> Quotes.jsx
