+++ src/Home/About.jsx
@@ -1,31 +1,12 @@
-import { useState, useEffect } from 'react'
 import useStyles from '../Theme/themes'
-
+import Quotes from './Quotes'
 import Box from '@mui/material/Box';
 
-// TODO: adjust profile pic and text size/alignment
+// DONE: move quotes logic into own comoponent --> Quotes.jsx
+// TODO: include my approach to programming
 
 const About = () => {
   const classes = useStyles();
-  const [quote,setQuote] = useState([])
-
-  const randomIdx = (min,max) => {
-    return Math.floor(Math.random()*(max-min+1)+min)
-  }
-  
-  useEffect(()=>{
-    const getProgrammerQuote = () => {
-      fetch(process.env.REACT_APP_QUOTES_API)
-        .then( res => {
-          if(!res.ok) console.error('err',res)
-          return res.json()
-        }).then(res => {
-          const idx = randomIdx(0,res.length)
-          setQuote([{...res[idx]}])}
-          )
-    }
-    getProgrammerQuote()
-  },[])
 
   return (
       <Box className={classes.aboutBackground}>
@@ -38,18 +19,7 @@ const About = () => {
               My willingness to work in a team is paramount to my conceptual knowledge and growth. 
               I want to be an effective and approachable developer that is always willing to teach those around me.
             </p>
-              {quote
-                  ? quote.map(item => (
-                    <div
-                      className={classes.aboutQuoteContainer}
-                      key={item.text}
-                    >
-                      <h3 className={classes.aboutQuote}>"{item.text}"</h3>
-                      <h5 className={classes.aboutQuote}>{`- ${item.author}`}</h5>
-                    </div>
-                  ))
-                  : null
-              }
+            <Quotes/>
           </div>
         </Box>
       </Box>
