+++ src/Home/CurrentEndeavors.jsx
@@ -1,0 +1,18 @@
+
+/* Actively working towards:
+
+- Current Skills Working On:
+    UDEMY: Understanding TypesScript, JS the Advanced Concepts, Modern HTML/CSS, Figma UI/UX Desgin Essentials
+- Current Projects Working on :
+    Drag and Drop (TypeScript)
+    Hotel WebSite, EdgeLedger (Modern HTML/CSS)
+    Personal Portfolio (Frontend: React, MUI, Git/GitHub)
+*/
+const CurrentEndeavors = () => {
+
+  return(
+    <>Current Endeavors component</>
+  )
+}
+
+export default CurrentEndeavors
